Fixed: Update Tracking areas only once when doing setFrame:
This speed improvement is very notable when resizing a browser window of altering the size of
split views.
The fix uses a variable on the window to control when to update the tracking areas.
If the view is not connected to a window the tracking areas will not be updated as they will
be when added to a window.