Update NibApplication's xib and cib to modern format
Any recent version of Xcode will complain that the original xib format is unsafe and force and upgrade when opening.
The MainMenu.xib is upgraded from the Xcode 3.2 format to a more modern one.
The corresponding MainMenu.cib is regenerated using nib2cib.

Xcode 9 and macOS 10.10 are used as the new baseline (minimum supported) platform requirements.